Some minor changes, getting the new architecture to run the Ultimate-II application. * SDcard timing corrected for Microblaze * USB hub, when it starts up with two or more devices attached; make sure to reset only one at a time * DMA load fixed, using a temporary RAM buffer * Tape load fixed, using a temporary RAM buffer

+++ software/filesystem/blockdev.cc
@@ -22,13 +22,13 @@ DSTATUS BlockDevice::status(void)
     return STA_NODISK;
 }
     
-DRESULT BlockDevice::read(BYTE *buffer, DWORD sector, BYTE count)
+DRESULT BlockDevice::read(BYTE *buffer, DWORD sector, int count)
 {
     return RES_NOTRDY;
 }
 
 #if	_READONLY == 0
-DRESULT BlockDevice::write(const BYTE *buffer, DWORD sector, BYTE count)
+DRESULT BlockDevice::write(const BYTE *buffer, DWORD sector, int count)
 {
     return RES_NOTRDY;
 }
