+++ software/filesystem/blockdev.h
@@ -62,9 +62,9 @@ public:
 
     virtual DSTATUS init(void);
     virtual DSTATUS status(void);
-    virtual DRESULT read(BYTE *, DWORD, BYTE);
+    virtual DRESULT read(BYTE *, DWORD, int);
 #if	_READONLY == 0
-    virtual DRESULT write(const BYTE *, DWORD, BYTE);
+    virtual DRESULT write(const BYTE *, DWORD, int);
 #endif
     virtual DRESULT ioctl(BYTE, void *);
